moving ban and timeout dialog
