+++ app/src/main/java/com/example/clicker/presentation/stream/StreamView.kt
@@ -179,7 +179,7 @@ fun StreamView(
                         openTimeoutDialog = {openTimeoutDialog.value = true},
                         timeoutDialogContent ={
                             if(openTimeoutDialog.value){
-                                TimeoutDialog(
+                                BottomModal.TimeoutDialog(
                                     onDismissRequest = {
                                         openTimeoutDialog.value = false
                                     },
@@ -210,7 +210,7 @@ fun StreamView(
                         openBanDialog = {openBanDialog.value = true},
                         banDialogContent ={
                             if(openBanDialog.value){
-                                BanDialog(
+                                BottomModal.BanDialog(
                                     onDismissRequest = {
                                         openBanDialog.value = false
                                     },
@@ -919,7 +919,7 @@ fun TextChat(
         }
 
 //        NoChatMode(modifier = Modifier.align(Alignment.Center))
-        EnterChat(
+        MainChat.TextChat.EnterChat(
             modifier = Modifier
                 .align(Alignment.BottomCenter)
                 .fillMaxWidth(),
@@ -1529,342 +1529,3 @@ fun ChatCard(
     }
 }
 
-
-
-@Composable
-fun EnterChat(
-    modifier: Modifier,
-    chat: (String) -> Unit,
-    modStatus: Boolean?,
-    filteredChatList: List<String>,
-    filterMethod: (String, String) -> Unit,
-    clickedAutoCompleteText: (String, String) -> String,
-    textFieldValue: MutableState<TextFieldValue>,
-    channelName: String?,
-    showModal: () -> Unit
-) {
-    // todo: I think we can move this to the viewModel
-
-    Column(modifier = modifier.background(androidx.compose.material3.MaterialTheme.colorScheme.primary)) {
-        FilteredMentionLazyRow(
-            filteredChatList =filteredChatList,
-            textFieldValue =textFieldValue,
-            clickedAutoCompleteText ={addedValue, currentValue ->clickedAutoCompleteText(addedValue,currentValue)}
-        )
-
-        TextFieldChat(
-            textFieldValue = textFieldValue,
-            modStatus = modStatus,
-            filterMethod = { username, text -> filterMethod(username, text) },
-            chat = { chatMessage -> chat(chatMessage) },
-            showModal = { showModal() }
-        )
-    }
-}
-
-@Composable
-fun FilteredMentionLazyRow(
-    filteredChatList: List<String>,
-    textFieldValue: MutableState<TextFieldValue>,
-    clickedAutoCompleteText: (String, String) -> String,
-){
-    LazyRow(modifier = Modifier.padding(vertical = 10.dp)) {
-        items(filteredChatList) {
-            Text(
-                it,
-                modifier = Modifier
-                    .padding(5.dp)
-                    .clickable {
-                        textFieldValue.value = TextFieldValue(
-                            text = clickedAutoCompleteText(textFieldValue.value.text, it),
-                            selection = TextRange(
-                                (textFieldValue.value.text + "$it ").length
-                            )
-                        )
-                    },
-                color = androidx.compose.material3.MaterialTheme.colorScheme.onPrimary
-            )
-        }
-    }
-}
-
-@Composable
-fun TextFieldChat(
-    textFieldValue: MutableState<TextFieldValue>,
-    modStatus: Boolean?,
-    filterMethod: (String, String) -> Unit,
-    chat: (String) -> Unit,
-    showModal: () -> Unit
-) {
-    val customTextSelectionColors = TextSelectionColors(
-        handleColor = androidx.compose.material3.MaterialTheme.colorScheme.secondary,
-        backgroundColor = androidx.compose.material3.MaterialTheme.colorScheme.secondary.copy(alpha = 0.5f)
-    )
-
-    Row(
-        modifier = Modifier.background(androidx.compose.material3.MaterialTheme.colorScheme.primary),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        if (modStatus != null && modStatus == true) {
-            AsyncImage(
-                model = "https://static-cdn.jtvnw.net/badges/v1/3267646d-33f0-4b17-b3df-f923a41db1d0/3",
-                contentDescription = stringResource(R.string.moderator_badge_icon_description)
-            )
-        }
-        CompositionLocalProvider(LocalTextSelectionColors provides customTextSelectionColors) {
-            TextField(
-
-                modifier = Modifier
-                    .weight(2f),
-                maxLines =1,
-                singleLine = true,
-                value = textFieldValue.value,
-                shape = RoundedCornerShape(8.dp),
-                onValueChange = { newText ->
-                    filterMethod("username", newText.text)
-                    textFieldValue.value = TextFieldValue(
-                        text = newText.text,
-                        selection = newText.selection
-                    )
-                    // text = newText
-                },
-                colors = TextFieldDefaults.textFieldColors(
-                    textColor = Color.White,
-                    backgroundColor = Color.DarkGray,
-                    cursorColor = Color.White,
-                    disabledLabelColor = Color.White,
-                    focusedIndicatorColor = Color.Transparent,
-                    unfocusedIndicatorColor = Color.Transparent
-                ),
-                placeholder = {
-                    Text(stringResource(R.string.send_a_message), color = Color.White)
-                }
-            )
-        }
-        if (textFieldValue.value.text.length > 0) {
-            Icon(
-                imageVector = Icons.Default.ArrowForward,
-                contentDescription = stringResource(R.string.send_chat),
-                modifier = Modifier
-                    .size(35.dp)
-                    .clickable { chat(textFieldValue.value.text) }
-                    .padding(start = 5.dp),
-                tint = androidx.compose.material3.MaterialTheme.colorScheme.onPrimary
-            )
-        } else {
-            Icon(
-                imageVector = Icons.Default.MoreVert,
-                contentDescription = stringResource(R.string.more_vert_icon_description),
-                modifier = Modifier
-                    .size(35.dp)
-                    .clickable { showModal() }
-                    .padding(start = 5.dp),
-                tint = androidx.compose.material3.MaterialTheme.colorScheme.onPrimary
-            )
-        }
-    }
-}
-
-@Composable
-fun TimeoutDialog(
-    onDismissRequest: () -> Unit,
-    username: String,
-
-    timeoutDuration: Int,
-    timeoutReason: String,
-    changeTimeoutDuration: (Int) -> Unit,
-    changeTimeoutReason: (String) -> Unit,
-    closeDialog: () -> Unit,
-    timeOutUser: () -> Unit
-) {
-    val secondary = androidx.compose.material3.MaterialTheme.colorScheme.secondary
-    val primary = androidx.compose.material3.MaterialTheme.colorScheme.primary
-    val onPrimary = androidx.compose.material3.MaterialTheme.colorScheme.onPrimary
-    val onSecondary = androidx.compose.material3.MaterialTheme.colorScheme.onSecondary
-    Dialog(onDismissRequest = { onDismissRequest() }) {
-        Card(
-            modifier = Modifier
-                .fillMaxWidth(),
-            shape = RoundedCornerShape(16.dp),
-            backgroundColor = primary,
-            border = BorderStroke(2.dp,secondary)
-        ) {
-            Column(
-                modifier = Modifier
-                    .padding(10.dp)
-                    .background(primary)
-            ) {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.SpaceAround
-                ) {
-                    Text(stringResource(R.string.timeout_text), fontSize = 22.sp,color = onPrimary)
-                    Text(username, fontSize = 22.sp,color = onPrimary)
-                }
-                Divider(color = secondary, thickness = 1.dp, modifier = Modifier.fillMaxWidth())
-                Text(stringResource(R.string.duration_text),color = onPrimary)
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceEvenly
-                ) {
-
-                    Column {
-                        RadioButton(
-                            colors =  RadioButtonDefaults.colors( selectedColor=secondary, unselectedColor = onPrimary),
-                            selected = timeoutDuration == 60,
-                            onClick = { changeTimeoutDuration(60) }
-                        )
-                        Text(stringResource(R.string.one_minute),color = onPrimary)
-                    }
-                    Column {
-                        RadioButton(
-                            colors =  RadioButtonDefaults.colors( selectedColor=secondary, unselectedColor = onPrimary),
-                            selected = timeoutDuration == 600,
-                            onClick = { changeTimeoutDuration(600) }
-                        )
-                        Text(stringResource(R.string.ten_minutes),color = onPrimary)
-                    }
-                    Column {
-                        RadioButton(
-                            colors =  RadioButtonDefaults.colors( selectedColor=secondary, unselectedColor = onPrimary),
-                            selected = timeoutDuration == 1800,
-                            onClick = { changeTimeoutDuration(1800) }
-                        )
-                        Text(stringResource(R.string.thirty_minutes),color = onPrimary)
-                    }
-                    Column {
-                        RadioButton(
-                            colors =  RadioButtonDefaults.colors( selectedColor=secondary, unselectedColor = onPrimary),
-                            selected = timeoutDuration == 604800,
-                            onClick = { changeTimeoutDuration(604800) }
-                        )
-                        Text(stringResource(R.string.one_week),color = onPrimary)
-                    }
-                }
-                OutlinedTextField(
-                    colors= TextFieldDefaults.textFieldColors(
-                        textColor = onPrimary, focusedLabelColor = onPrimary,
-                        focusedIndicatorColor = onPrimary, unfocusedIndicatorColor = onPrimary, unfocusedLabelColor = onPrimary),
-                    value = timeoutReason,
-                    onValueChange = { changeTimeoutReason(it) },
-                    label = { Text(stringResource(R.string.reason)) }
-                )
-                Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.End) {
-                    Button(
-                        onClick = { onDismissRequest() },
-                        modifier = Modifier.padding(10.dp),
-                        colors = ButtonDefaults.buttonColors(backgroundColor = secondary)
-                    ) {
-                        Text(stringResource(R.string.cancel),color = onSecondary)
-                    }
-                    // todo: Implement the details of the timeout implementation
-                    Button(
-                        colors = ButtonDefaults.buttonColors(backgroundColor = secondary),
-                        onClick = {
-                        closeDialog()
-                        timeOutUser()
-                    }, modifier = Modifier.padding(10.dp)) {
-                        Text(stringResource(R.string.timeout_confirm),color = onSecondary)
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun BanDialog(
-    onDismissRequest: () -> Unit,
-    username: String,
-    banDuration: Int,
-    banReason: String,
-    changeBanDuration: (Int) -> Unit,
-    changeBanReason: (String) -> Unit,
-    banUser: (BanUser) -> Unit,
-    clickedUserId: String,
-    closeDialog: () -> Unit,
-    closeBottomModal: () -> Unit
-) {
-    val secondary = androidx.compose.material3.MaterialTheme.colorScheme.secondary
-    val primary = androidx.compose.material3.MaterialTheme.colorScheme.primary
-    val onPrimary = androidx.compose.material3.MaterialTheme.colorScheme.onPrimary
-    val onSecondary = androidx.compose.material3.MaterialTheme.colorScheme.onSecondary
-    Dialog(
-        onDismissRequest = { onDismissRequest() },
-    ) {
-        Card(
-            modifier = Modifier
-                .fillMaxWidth(),
-            shape = RoundedCornerShape(16.dp),
-            backgroundColor = primary,
-            border = BorderStroke(2.dp,secondary)
-        ) {
-            Column(
-                modifier = Modifier
-                    .padding(10.dp)
-                    .background(primary)
-            ) {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.Start
-                ) {
-                    Text(stringResource(R.string.ban), fontSize = 22.sp,color = onPrimary)
-                    Text(username, fontSize = 22.sp,color = onPrimary)
-                }
-                Divider(color = androidx.compose.material3.MaterialTheme.colorScheme.secondary, thickness = 1.dp, modifier = Modifier.fillMaxWidth())
-                Text(stringResource(R.string.duration_text),color = onPrimary)
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.Start
-                ) {
-
-                    Column {
-                        RadioButton(
-                            colors =  RadioButtonDefaults.colors( selectedColor=secondary, unselectedColor = onPrimary),
-                            selected = banDuration == 0,
-                            onClick = { changeBanDuration(0) }
-                        )
-                        Text(stringResource(R.string.permanently),color = onPrimary)
-                    }
-                }
-                OutlinedTextField(
-                    colors= TextFieldDefaults.textFieldColors(
-                        textColor = onPrimary, focusedLabelColor = onPrimary,
-                        focusedIndicatorColor = onPrimary, unfocusedIndicatorColor = onPrimary, unfocusedLabelColor = onPrimary),
-                    value = banReason,
-                    onValueChange = { changeBanReason(it) },
-                    label = { Text(stringResource(R.string.reason),color = onPrimary) }
-                )
-                Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.End) {
-                    Button(
-                        colors = ButtonDefaults.buttonColors(backgroundColor = secondary),
-                        onClick = { onDismissRequest() }, modifier = Modifier.padding(10.dp)
-                    ) {
-                        Text(stringResource(R.string.cancel),color =onSecondary)
-                    }
-
-                    Button(
-                        colors = ButtonDefaults.buttonColors(backgroundColor = secondary),
-                        onClick = {
-                            closeDialog()
-                            closeBottomModal()
-                            banUser(
-                                BanUser(
-                                    data = BanUserData(
-                                        user_id = clickedUserId,
-                                        reason = banReason
-                                    )
-                                )
-                            )
-                        },
-                        modifier = Modifier.padding(10.dp)
-                    ) {
-                        Text(stringResource(R.string.ban),color =onSecondary)
-                    }
-                }
-            }
-        }
-    }
-}
